refactor: move CLI workflow output resolution to runtime
